Garbage styles removed
I have removed all the paragraph and list styles prefixed 'WW' in all four Parts.

In Part 3:

The garbage attribute value ‘linearpolynomialexponential’ results from the way the extraction script was interpreting the content of the final paragraph of 20.45 chart:regression-intercept-value, which has the paragraph style ‘Default Value’ applied, but contains three attribute values that are each the “default” in different contexts. I have therefore re-styled that paragraph ‘Text Body’, so that it won’t be processed by the extraction script.

Alfred found duplicate citations for RFC 6838 in 1.3 Normative References. I have deleted the first instance. Alfred has checked the other editorial issues that he had raised on Jira and they are all fixed.

I have looked at how the styling of ODF 1.3 CSD02 compares with ODF 1.2 OS, and the main difference, apart from a change of font from ‘Arial’ to ‘Liberation Sans’, is that in ODF 1.2 OS the paragraph style ‘Text Body’ has slightly more space above and below: 0.14 cm compared with 0.10 cm in ODF 1.3 CSD02. I feel inclined to leave it be, as I’m not sure that it will improve the appearance noticeably, but it will increase the number of pages.

In passing I noticed a few cases where a space is missing following punctuation marks. These have been fixed in the following cases:

-	9.3.3, first paragraph: space missing before final sentence.
-	19.242.1, first paragraph: space missing before third sentence.
-	19.628, Step 2: period following the colon should be a space.
-	20.44, first paragraph: space missing before the second sentence.
-	Appendix E, heading: space missing following first comma.