Saving latest updates as WD 02
In WD 02 there are minor updates to Parts 1 and 4. In Parts 2 and 3 the generated text has been updated from the current drafts of the RNG schemas.

NOTE: The master tracked-changes versions of Parts 2 and 3 do NOT contain updated generated text, as this can only be added after accepting all tracked changes.