Re-organizing documents and schemas for formal stages
The structure of /odf-tc/docs/odf1.4 has been re-organized to reflect the way that the OASIS document server stores documents in the formal stages.

The documents that the ODF TC has been working on up to this point, including the master tracked documents and Working Drafts 01 and 02 remain where there were previously, i.e. in folders 'part1-introduction' etc directly within 'odf-tc/docs/odf1.4'.

The documents for the formal approval stages will have their own folders within 'odf1.4', e.g. 'csd01', and the structure below that reflects the structure in which documents are expected to be held in the OASIS document server.

The latest drafts of Parts 1 through 4 in the formal stages are also available directly from the top-level folder 'odf-fc/docs/odf1.4', but with the stage identifier (e.g. 'csd01') removed from the filename in each case. This reflects current practice on the OASIS document server, as is documented in the front matter of each Part.