Removal of editorial comment in Part 4, 6.5.1 General
A comment had been added at the WD stage to a changed graphic in 6.5.1, but had not been removed from the CSD, CS or OS. While unobtrusive in the .ODT file and omitted from the .PDF file, the comment appeared as yellow-highlighted text in the .HTML file. This has now been removed in the .ODT file and the .HTML file.